refactor(app): move the settings popup state into ainb-app
ConfigPopupState, its value and type enums and their editing logic move to
ainb_app::components::config_popup; the popup renderer stays in ainb-core.
