+++ ainb-tui/crates/ainb-app/src/components/mod.rs
@@ -2,6 +2,7 @@
 // their impls and the reducers that do not draw. The draw functions stay in
 // `ainb-core::components`, which re-exports this module.
 
+pub mod config_popup;
 pub mod onboarding;
 pub mod setup_menu;
 pub mod skills;
